+++ greeclimate/device.py
@@ -5,7 +5,7 @@ import re
 import typing
 from asyncio import AbstractEventLoop
 from enum import IntEnum, unique
-from typing import Union
+from typing import Union, Optional, Any
 
 from greeclimate.cipher import CipherV1, CipherV2
 from greeclimate.deviceinfo import DeviceInfo
@@ -180,6 +180,9 @@ class Device(DeviceProtocol2, Taskable):
         self._properties = {}
         self._dirty = []
 
+        self._valid_state: asyncio.Event = asyncio.Event()
+        self._valid_state.clear()
+
     async def bind(
         self,
         key: str = None,
@@ -245,12 +248,14 @@ class Device(DeviceProtocol2, Taskable):
     async def __bind_internal(self, cipher: Union[CipherV1, CipherV2]):
         """Internal binding procedure, do not call directly"""
         await self.send(self.create_bind_message(self.device_info), cipher=cipher)
-        task = asyncio.create_task(self.ready.wait())
+        task = self._loop.create_task(self.ready.wait())
         await asyncio.wait_for(task, timeout=self._bind_timeout)
 
     def handle_device_bound(self, key: str) -> None:
         """Handle the device bound message from the device"""
+        DeviceProtocol2.handle_device_bound(self, key)
         self.device_cipher.key = key
+        self._loop.create_task(self.update_state())
 
     async def request_version(self) -> None:
         """Request the firmware version from the device."""
@@ -263,11 +268,8 @@ class Device(DeviceProtocol2, Taskable):
         except asyncio.TimeoutError:
             raise DeviceTimeoutError
 
-    async def update_state(self, wait_for: float = 30):
+    async def update_state(self):
         """Update the internal state of the device structure of the physical device, 0 for no wait
-
-        Args:
-            wait_for (object): How long to wait for an update from the device
         """
         if not self.device_cipher:
             await self.bind()
@@ -305,11 +307,10 @@ class Device(DeviceProtocol2, Taskable):
                 self._logger.info(f"Device version changed to {self.version}, hid {self.hid}")
             self._logger.debug(f"Using device temperature {self.current_temperature}")
 
-    async def push_state_update(self, wait_for: float = 30):
-        """Push any pending state updates to the unit
+        self._valid_state.set()
 
-        Args:
-            wait_for (object): How long to wait for an update from the device, 0 for no wait
+    async def push_state_update(self):
+        """Push any pending state updates to the unit
         """
         if not self._dirty:
             return
@@ -330,13 +331,14 @@ class Device(DeviceProtocol2, Taskable):
                     Props.TEMP_UNIT.value
                 )
 
-        self._dirty.clear()
-
         try:
             await self.send(self.create_command_message(self.device_info, **props))
 
         except asyncio.TimeoutError:
             raise DeviceTimeoutError
+        else:
+            self._dirty.clear()
+
 
     def __eq__(self, other):
         """Compare two devices for equality based on their properties state and device info."""
@@ -351,7 +353,7 @@ class Device(DeviceProtocol2, Taskable):
     def raw_properties(self) -> dict:
         return self._properties
 
-    def get_property(self, name):
+    def get_property(self, name) -> Optional[Any]:
         """Generic lookup of properties tracked from the physical device"""
         if self._properties:
             return self._properties.get(name.value)
@@ -370,15 +372,20 @@ class Device(DeviceProtocol2, Taskable):
                 self._dirty.append(name.value)
 
     @property
-    def power(self) -> bool:
-        return bool(self.get_property(Props.POWER))
+    def has_valid_state(self) -> bool:
+        return self._valid_state.is_set()
+
+    @property
+    def power(self) -> Optional[bool]:
+        prop = self.get_property(Props.POWER)
+        return bool(prop) if prop is not None else None
 
     @power.setter
     def power(self, value: int):
         self.set_property(Props.POWER, int(value))
 
     @property
-    def mode(self) -> int:
+    def mode(self) -> Optional[int]:
         return self.get_property(Props.MODE)
 
     @mode.setter
@@ -402,9 +409,11 @@ class Device(DeviceProtocol2, Taskable):
         return f["f"]
 
     @property
-    def target_temperature(self) -> int:
+    def target_temperature(self) -> Optional[int]:
         temset = self.get_property(Props.TEMP_SET)
         temrec = self.get_property(Props.TEMP_BIT)
+        if temset is None or temrec is None:
+            return None
         return self._convert_to_units(temset, temrec)
 
     @target_temperature.setter
@@ -423,7 +432,7 @@ class Device(DeviceProtocol2, Taskable):
             self.set_property(Props.TEMP_SET, int(value))
 
     @property
-    def temperature_units(self) -> int:
+    def temperature_units(self) -> Optional[int]:
         return self.get_property(Props.TEMP_UNIT)
 
     @temperature_units.setter
@@ -431,10 +440,11 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.TEMP_UNIT, int(value))
 
     @property
-    def current_temperature(self) -> int:
+    def current_temperature(self) -> Optional[int]:
         prop = self.get_property(Props.TEMP_SENSOR)
         bit = self.get_property(Props.TEMP_BIT)
         if prop is not None:
+            bit = bit if bit is not None else 0
             v = self.version and int(self.version.split(".")[0])
             try:
                 if v == 4:
@@ -447,7 +457,7 @@ class Device(DeviceProtocol2, Taskable):
         return self.target_temperature
 
     @property
-    def fan_speed(self) -> int:
+    def fan_speed(self) -> Optional[int]:
         return self.get_property(Props.FAN_SPEED)
 
     @fan_speed.setter
@@ -455,32 +465,36 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.FAN_SPEED, int(value))
 
     @property
-    def fresh_air(self) -> bool:
-        return bool(self.get_property(Props.FRESH_AIR))
+    def fresh_air(self) -> Optional[bool]:
+        prop = self.get_property(Props.FRESH_AIR)
+        return bool(prop) if prop is not None else None
 
     @fresh_air.setter
     def fresh_air(self, value: bool):
         self.set_property(Props.FRESH_AIR, int(value))
 
     @property
-    def xfan(self) -> bool:
-        return bool(self.get_property(Props.XFAN))
+    def xfan(self) -> Optional[bool]:
+        prop = self.get_property(Props.XFAN)
+        return bool(prop) if prop is not None else None
 
     @xfan.setter
     def xfan(self, value: bool):
         self.set_property(Props.XFAN, int(value))
 
     @property
-    def anion(self) -> bool:
-        return bool(self.get_property(Props.ANION))
+    def anion(self) -> Optional[bool]:
+        prop = self.get_property(Props.ANION)
+        return bool(prop) if prop is not None else None
 
     @anion.setter
     def anion(self, value: bool):
         self.set_property(Props.ANION, int(value))
 
     @property
-    def sleep(self) -> bool:
-        return bool(self.get_property(Props.SLEEP))
+    def sleep(self) -> Optional[bool]:
+        prop = self.get_property(Props.SLEEP)
+        return bool(prop) if prop is not None else None
 
     @sleep.setter
     def sleep(self, value: bool):
@@ -488,15 +502,16 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.SLEEP_MODE, int(value))
 
     @property
-    def light(self) -> bool:
-        return bool(self.get_property(Props.LIGHT))
+    def light(self) -> Optional[bool]:
+        prop = self.get_property(Props.LIGHT)
+        return bool(prop) if prop is not None else None
 
     @light.setter
     def light(self, value: bool):
         self.set_property(Props.LIGHT, int(value))
 
     @property
-    def horizontal_swing(self) -> int:
+    def horizontal_swing(self) -> Optional[int]:
         return self.get_property(Props.SWING_HORIZ)
 
     @horizontal_swing.setter
@@ -504,7 +519,7 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.SWING_HORIZ, int(value))
 
     @property
-    def vertical_swing(self) -> int:
+    def vertical_swing(self) -> Optional[int]:
         return self.get_property(Props.SWING_VERT)
 
     @vertical_swing.setter
@@ -512,7 +527,7 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.SWING_VERT, int(value))
 
     @property
-    def quiet(self) -> bool:
+    def quiet(self) -> Optional[int]:
         return self.get_property(Props.QUIET)
 
     @quiet.setter
@@ -520,33 +535,36 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.QUIET, 2 if value else 0)
 
     @property
-    def turbo(self) -> bool:
-        return bool(self.get_property(Props.TURBO))
+    def turbo(self) -> Optional[bool]:
+        prop = self.get_property(Props.TURBO)
+        return bool(prop) if prop is not None else None
 
     @turbo.setter
     def turbo(self, value: bool):
         self.set_property(Props.TURBO, int(value))
 
     @property
-    def steady_heat(self) -> bool:
-        return bool(self.get_property(Props.STEADY_HEAT))
+    def steady_heat(self) -> Optional[bool]:
+        prop = self.get_property(Props.STEADY_HEAT)
+        return bool(prop) if prop is not None else None
 
     @steady_heat.setter
     def steady_heat(self, value: bool):
         self.set_property(Props.STEADY_HEAT, int(value))
 
     @property
-    def power_save(self) -> bool:
-        return bool(self.get_property(Props.POWER_SAVE))
+    def power_save(self) -> Optional[bool]:
+        prop = self.get_property(Props.POWER_SAVE)
+        return bool(prop) if prop is not None else None
 
     @power_save.setter
     def power_save(self, value: bool):
         self.set_property(Props.POWER_SAVE, int(value))
 
     @property
-    def target_humidity(self) -> typing.Optional[int]:
-        value = self.get_property(Props.HUM_SET)
-        return (15 + value * 5) if value is not None else None
+    def target_humidity(self) -> Optional[int]:
+        prop = self.get_property(Props.HUM_SET)
+        return 15 + (prop * 5) if prop is not None else None
 
     @target_humidity.setter
     def target_humidity(self, value: int):
@@ -557,17 +575,19 @@ class Device(DeviceProtocol2, Taskable):
         self.set_property(Props.HUM_SET, (value - 15) // 5)
 
     @property
-    def dehumidifier_mode(self):
+    def dehumidifier_mode(self) -> Optional[int]:
         return self.get_property(Props.DEHUMIDIFIER_MODE)
 
     @property
-    def current_humidity(self) -> int:
+    def current_humidity(self) -> Optional[int]:
         return self.get_property(Props.HUM_SENSOR)
 
     @property
-    def clean_filter(self) -> bool:
-        return bool(self.get_property(Props.CLEAN_FILTER))
+    def clean_filter(self) -> Optional[bool]:
+        prop = self.get_property(Props.CLEAN_FILTER)
+        return bool(prop) if prop is not None else None
 
     @property
-    def water_full(self) -> bool:
-        return bool(self.get_property(Props.WATER_FULL))
+    def water_full(self) -> Optional[bool]:
+        prop = self.get_property(Props.WATER_FULL)
+        return bool(prop) if prop is not None else None